refactor: added lib.rs, updated: imports, added tests for regression module
