+++ crates/bench/src/lib.rs
@@ -1,0 +1,11 @@
+pub mod allocator;
+pub mod benchmark;
+pub mod cli;
+pub mod io;
+pub mod manifest;
+pub mod markdown;
+pub mod profiler;
+pub mod promotion;
+pub mod regression;
+pub mod report;
+pub mod utils;
